pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "healthcare-medallion"
version = "0.1.0"
description = "Healthcare data engineering starter project using Python, Spark, SQL, and medallion architecture."
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "delta-spark>=3.2.0,<4.0.0",
    "PyYAML>=6.0.1,<7.0.0",
    "pyspark>=3.5.0,<4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9.0.0",
]
airflow = [
    "apache-airflow>=2.9.0,<3.0.0; platform_system != 'Windows'",
]

[project.scripts]
healthcare-medallion = "healthcare_medallion.pipeline:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
